+++ provider/baked/src/export.rs
@@ -622,7 +622,7 @@ impl DataExporter for BakedExporter {
                             #[allow(dead_code)]
                             pub(crate) const MUST_USE_MAKE_PROVIDER_MACRO: () = ();
                         }
-                        icu_provider::impl_data_provider_never_marker!($name);
+                        icu_provider::marker::impl_data_provider_never_marker!($name);
                     };
                 }
                 #[doc(inline)]
