+++ provider/blob/src/export/blob_exporter.rs
@@ -7,7 +7,7 @@
 
 use crate::blob_schema::*;
 use icu_provider::datagen::*;
-use icu_provider::{prelude::*, DataMarkerPathHash};
+use icu_provider::{marker::DataMarkerPathHash, prelude::*};
 use std::collections::{BTreeMap, BTreeSet, HashMap};
 use std::sync::Mutex;
 use zerotrie::ZeroTrieSimpleAscii;
