+++ provider/core/macros/src/lib.rs
@@ -373,7 +373,7 @@ fn data_struct_impl(attr: DataStructArgs, input: DeriveInput) -> TokenStream2 {
             result.extend(quote!(
                 impl icu_provider::DataMarker for #marker_name {
                     const INFO: icu_provider::DataMarkerInfo = {
-                        let mut info = icu_provider::DataMarkerInfo::from_path(icu_provider::data_marker_path!(#path_str));
+                        let mut info = icu_provider::DataMarkerInfo::from_path(icu_provider::marker::data_marker_path!(#path_str));
                         info.is_singleton = #singleton;
                         info.fallback_config.priority = #fallback_by_expr;
                         info.fallback_config.extension_key = #extension_key_expr;
