+++ provider/datagen/tests/test-options.rs
@@ -13,7 +13,6 @@ use icu_datagen::prelude::*;
 use icu_datagen::FallbackOptions;
 use icu_provider::datagen::*;
 use icu_provider::hello_world::*;
-use icu_provider::make_exportable_provider;
 use icu_provider::prelude::*;
 use testutil::TestingExporter;
