+++ provider/fs/tests/test_file_io.rs
@@ -75,7 +75,7 @@ fn test_errors() {
         }
         impl DataMarker for WrongV1Marker {
             const INFO: DataMarkerInfo =
-                DataMarkerInfo::from_path(icu_provider::data_marker_path!("nope@1"));
+                DataMarkerInfo::from_path(icu_provider::marker::data_marker_path!("nope@1"));
         }
 
         let err: Result<DataResponse<WrongV1Marker>, DataError> =
